agents/email_drafter_agent/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
google-api-python-client>=2.140.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-cloud-firestore>=2.19.0
slack-sdk>=3.33.0
